LayoutTests/crypto/subtle/aes-cbc-256-encrypt-decrypt-expected.txt

Test AES-CBC encrypt and decrypt functions with a 256 bit key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a raw AES key from string literal...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'aes-cbc'
PASS key.algorithm.length is 256
PASS key.usages is ['encrypt', 'decrypt']
Using the key to encrypt plaintext...
PASS byteArrayToHexString(new Uint8Array(encryptionResult)) is '[f5 8c 4c 04 d6 e5 f1 ba 77 9e ab fb 5f 7b fb d6 9c fc 4e 96 7e db 80 8d 67 9f 77 7b c6 70 2c 7d 39 f2 33 69 a9 d9 ba cf a5 30 e2 63 04 23 14 61 b2 eb 05 e2 c3 9b e9 fc da 6c 19 07 8c 6a 9d 1b 3f 46 17 96 d6 b0 d6 b2 e0 c2 a7 2b 4d 80 e6 44]'
Decrypting it back...
PASS new Uint8Array(decryptionResult) is plaintext
PASS successfullyParsed is true

TEST COMPLETE
